backend/requirements.txt
Flask==2.2.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
PyMySQL==1.0.3
python-dotenv==1.0.0
Werkzeug==2.2.3
marshmallow==3.19.0
Flask-Cors==3.0.10
argon2-cffi==21.3.0
nltk==3.8.1
PyPDF2==3.0.1
python-docx==0.8.11
pytest==7.3.1  